More granular CheckConnection reporting. (#2485)
A CheckConnection job can succeed and have the check fail. Because of this, just tracking the job_attempt does not let us distinguish between the two cases 1) Job succeeds and Check fails 2) Job succeeds and Check succeeds.

Add a check_connection_outcome field in the metadata that pulls from the StandardCheckConnectionOutput.getStatus field.